requirements.txt
discord.py[voice]==2.3.2
pdf2image==1.16.2